pyproject.toml
[tool.poetry]
name = "azure-app-to-get-image-details-from-acr"
version = "0.1.0"
description = ""
authors = ["githubofkrishnadhas <krishnadhasnk1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-identity = "^1.17.1"
azure-containerregistry = "^1.2.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
argparse = "^1.4.0"
streamlit = "^1.36.0"
azure-mgmt-resource = "^23.1.1"
azure-mgmt-resourcegraph = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
